refactor: rename `remote_file_uploader` to `gcp_file_uploader`
